Add check that operator split only used with bdfOrder = 1
Current implementation does not support higher order schemes, so be
sure user doesn't run that way.  NB: Required adding the order to the
temporalSchemeCoefficients struct b/c otherwise the model classes (in
particular ReactingFlow here) only know the coefficients, not the
order.  The coefficients cannot be used for this check b/c the first
step is always a BDF1 step.