Minor fixes to thermal conductivity
NB: Also modified test/inputs/input.reactTable.ini to set sponge
multiplier to 1 (previously was 99000).  Without this, the
reactFlow-table.test case doesn't run, I suspect because the
relationship between the unsteady and diffusive terms in the energy
equation is such that the diffusive term dominates so strongly that,
at the tolerance we have, the linear solve leads to inaccuracy in the
time evolution.  I have not investigated systematically though.