Fixes for ambipolar diffusion on low Mach path
In splitting out the diffusivity calculations for the low Mach path,
it appears that the ambipolar modification was lost.  This commit
addresses this issue by implementing the ambipolar modification to the
charged species diffusivities in
ArgonMixtureTransport::computeMixtureAverageDiffusivity

For context, for the compressible solver, the diffusive fluxes are
evaluated by computing the diffusion velocity in
ArgonMixtureTransport::ComputeFluxMolecularTransport, which properly
accounts for the ambipolar diffusion effect.  However, the low Mach
solver uses ArgonMixtureTransport::computeMixtureAverageDiffusivity,
which, until now, has returned only the mixture averaged
diffusivities, without accounting for the ambipolar effect.

This commit represents a minimal fix, which is valid for the case with
a single ion species, which is all we need now.  If multiple ions are
present, the ambipolar modification is more complicated and cannot be
represented via a simple modification of the ion and electron
diffusivities, because the effect couples the diffusion of one ion to
the other ions so that there are 'off-diagonal' terms.  However, we do
not need this level of generality yet.