Bug fix for enforcing quasi-neutrality on low Mach path
Was using an incorrectly sized temporary storage Vector.  This did not
affect the 3 species case (b/c the bug is that the stride between
species was too large, which is irrelevant with only 1 active specie)
so no existing test cases are affected.  Was found while attempting to
run 6 species torch case.