Towards a nonlinear implicit step for chemistry
This commit begins implementation of a nonlinear, implicit solve for
the thermochemical evolution (i.e., just the local in space portion of
the energy and species equations).  The eventual goal is to enable
implicit time marching for this part of the system to avoid the very
small time steps that are required in the current method (where the
reaction terms are operator split from convection-diffusion but
marched explicitly at their own time step) when fast reactions are
present.

To enable this development, we have to rearrange the code to enable
evaluation of the chemical source terms at each point in space, rather
than for the full field, so that we can update each point in space
independently.  This commit implements the calculations necessary to
evaluate the residual, but does not implement the solve.