deceptively large push, mostly to make tps compatible with c++17 and mfem 4.8.
This primary change is to rename the tps RiemannSolver class to
RiemannSolverTPS.  This is done b/c mfem has introduced a
RiemannSolver into the mfem namespace, and tps is very promiscuous
about using namespace mfem, leading to a clash.  We could (probably
should) fix that, but renaming the tps RiemannSolver is quicker.