Use #if to get const-correctness pre- and post- mfem v4.7 change
In mfem v4.7 the ElementTransformation::mesh data member was changed
to const.  We use this in the
GradientVectorGridFunctionCoefficient::Eval function, which reproduces
some code from mfem::RefinedToCoarse (see mfem/fem/coefficient.cpp).
So, for mfem 4.7 and later we require const here.

But, prior to mfem v4.7, we cannot use const here b/c the subsequent
function call to mesh::GetRefinementTransforms() violates
const-correctness.

So, the solution is the #if in this commit.