Update gll rules in LteThermoChem
so that the quad points correspond to the nodes.  This leads to a
diagonal mass matrix and ensures that as long as the solution remains
physical at the nodes, it is physical at the quad points.  But, it
can't always be used (e.g., will fail for axisymmetric and for some
grids) so should be used with caution.